Refactor Link in Bio processing step
