+++ client/landing/stepper/declarative-flow/internals/steps-repository/index.ts
@@ -37,7 +37,6 @@ export { default as linkInBioSetup } from './link-in-bio-setup';
 export { default as completingPurchase } from './completing-purchase';
 export { default as chooseADomain } from './choose-a-domain';
 export { default as launchpad } from './launchpad';
-export { default as linkInBioProcessingStep } from './link-in-bio-processing-step';
 export { default as processingFake } from './processing-fake';
 export { default as subscribers } from './subscribers';
 
@@ -81,6 +80,5 @@ export type StepPath =
 	| 'completingPurchase'
 	| 'intro'
 	| 'launchpad'
-	| 'linkInBioProcessingStep'
 	| 'processingFake'
 	| 'subscribers';
