--- client/landing/stepper/declarative-flow/internals/steps-repository/link-in-bio-processing-step/styles.scss
@@ -1,60 +1,0 @@
-@import '@automattic/onboarding/styles/mixins';
-
-.link-in-bio-processing-step {
-
-	.image-right {
-		background-image:  url( 'calypso/assets/images/onboarding/link-in-bio-loader-right.png' );
-
-		width: 142px;
-		height: 220px;
-		position: fixed;
-		right: -78px;
-		top: 60px;
-		border-radius: 170px; /* stylelint-disable-line scales/radii */
-		background-size: 50%;
-
-		@include break-medium {
-			width: 340px;
-			height: 522px;
-			right: -170px;
-		}
-	}
-	.image-left {
-		background-image:  url( 'calypso/assets/images/onboarding/link-in-bio-loader-left.png' );
-
-		width: 246px;
-		height: 220px;
-		position: fixed;
-		border-radius: 66px; /* stylelint-disable-line scales/radii */
-		left: -168px;
-		background-size: 34%;
-		bottom: 10%;
-
-		@include break-medium {
-			width: 336px;
-			height: 522px;
-			background-size: 50%;
-			border-radius: 170px;  /* stylelint-disable-line scales/radii */
-		}
-		
-	}
-	
-	background: #d0cce3;
-	position: fixed;
-	width: 100%;
-	height: 100%;
-
-	.step-container__content {
-		display: flex;
-		justify-content: center;
-		width: 100%;
-		
-		.processing-step {
-			width: 100%;
-		}
-	}
-	
-	.progress-bar {
-		display: none;
-	}
-}
